backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Compress==1.14
Flask-RESTful==0.3.10
python-dotenv==1.0.0
Werkzeug==3.0.1
fastapi==0.109.0
uvicorn==0.27.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.22.4,<2.0.0
pandas==2.1.4
scipy>=1.8.0
joblib>=1.2.0

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
gensim==4.3.2

# Data Processing
python-magic==0.4.27
pillow==10.1.0
tqdm==4.66.1

# ===== Multi-Modal Support =====

# Image Processing & OCR
pytesseract==0.3.10
easyocr>=1.7.0
opencv-python-headless>=4.9.0.80

# Deep Learning for Vision (NEW)
torch>=2.0.0
transformers>=4.35.0
ultralytics>=8.0.0

# Audio Processing & Speech-to-Text
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
pydub>=0.25.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
webrtcvad>=2.0.10  # Voice Activity Detection (optional but recommended)

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# API & Utilities
requests==2.31.0
python-multipart==0.0.6
pyyaml==6.0.1
jsonschema==4.20.0
orjson==3.9.10

# Development & Testing
black==23.12.1
pylint==3.0.3
pytest==7.4.4
pytest-cov==4.1.0
requests==2.31.0  # For API testing
